requirements.txt
python-telegram-bot>=12.0.0b1
pyowm>=2.10.0
python-dotenv>=0.10.1